use TestCafe t.report API
