+++ src/fixture.ts
@@ -98,6 +98,7 @@ export class Fixture {
     test: Test,
     screenshotAssets: Assets,
     videoAssets: Assets,
+    artifacts: Assets,
   ) {
     if (!this.browserTestRuns.has(userAgent)) {
       this.browserTestRuns.set(userAgent, new BrowserTestRun(userAgent));
@@ -121,7 +122,7 @@ export class Fixture {
     }
 
     tr.addTest(this.path, this.name, test);
-    tr.addAssets([...screenshotAssets, ...videoAssets]);
+    tr.addAssets([...screenshotAssets, ...videoAssets, ...artifacts]);
   }
 
   get testRuns() {
