+++ src/index.js
@@ -83,7 +83,7 @@ module.exports = function () {
     },
 
     reportTestDone: async function (name, testRunInfo, meta) {
-      this.sauceJsonReporter.reportTestDone(name, testRunInfo, meta);
+      await this.sauceJsonReporter.reportTestDone(name, testRunInfo, meta);
 
       if (this.disableUpload) {
         return;
@@ -195,10 +195,6 @@ module.exports = function () {
               assets: browserTestRun.assets,
               userAgent: browserTestRun.userAgent,
             };
-            const artifacts = await this.reporter.collectArtifacts(
-              path.basename(fixture.path),
-            );
-            session.assets.push(...artifacts);
             try {
               const job = await this.reporter.reportSession(session);
               this.setIndent(this.indentWidth * 4)
@@ -288,7 +284,12 @@ module.exports = function () {
 
       this.write(title, name, testRunInfo, meta);
 
-      this._renderReportData(testRunInfo.reportData, name, testRunInfo, meta);
+      this._renderReportData(
+        testRunInfo.reportData,
+        testRunInfo.browsers,
+        testRunInfo,
+        meta,
+      );
 
       if (hasErr) {
         this._renderErrors(testRunInfo.errs, name, testRunInfo, meta);
@@ -332,7 +333,7 @@ module.exports = function () {
     },
 
     _renderReportData(reportData, browsers, name, testRunInfo, meta) {
-      if (!reportData) return;
+      if (!reportData || !browsers) return;
 
       if (!Object.values(reportData).some((data) => data.length)) return;
 
@@ -355,7 +356,7 @@ module.exports = function () {
         browserReportData.forEach((data) => {
           this.setIndent(this.indentWidth * dataIndent)
             .newline()
-            .write(`- ${data}`, name, testRunInfo, meta);
+            .write(`- ${JSON.stringify(data)}`, name, testRunInfo, meta);
         });
       });
     },
